# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GuimaAPI",
        "description": "API para gerenciamento de alunos, treino, dieta, aulas e avaliações físicas.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Listar todos os usuários",
                "operationId": "3ab81e3f9b24b4f36ab08d72349ac632",
                "responses": {
                    "200": {
                        "description": "Retorna a lista de usuários"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Registrar um novo usuário",
                "description": "A senha padrão é o cpf do usuário.",
                "operationId": "ba548b97148bb5c0d8cab3bc3f128bfd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Joao"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joao@email.com"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "01-01-2003"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "111.111.111-13"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Rua teste"
                                    },
                                    "profile_type": {
                                        "type": "string",
                                        "example": "aluno"
                                    },
                                    "plano": {
                                        "type": "string",
                                        "example": "anual"
                                    },
                                    "vencimento": {
                                        "type": "string",
                                        "example": "2025-12-12"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o usuário registrado"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Mostrar um usuário específico por id",
                "operationId": "daaf7a6169ee9348c3fddb07778598ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buscar por id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o usuário específico por id"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Deletar um usuário específico por id",
                "description": "Deletar um usuário específico por id",
                "operationId": "c30c64b8b7730b65c7714c9af4a6f5fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do usuário",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuário deletado com sucesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Atualizar um usuário",
                "description": "Para atualizar a senha, utilize o endpoint POST /api/v1/auth/reset-password.",
                "operationId": "12801392b8a73f66c3932bc46754d201",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Joao"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joao@email.com"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "01-01-2003"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "111.111.111-13"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Rua teste"
                                    },
                                    "profile_type": {
                                        "type": "string",
                                        "example": "aluno"
                                    },
                                    "plano": {
                                        "type": "string",
                                        "example": "anual"
                                    },
                                    "vencimento": {
                                        "type": "string",
                                        "example": "2025-12-12"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o usuário atualizado"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}